refactor(config): split write preparation primitives
