+++ src/config/io.write-prepare.ts
@@ -13,14 +13,10 @@ import { normalizeAgentId } from "../routing/session-key.js";
 import { parseConfigPathArrayIndex } from "../shared/path-array-index.js";
 import { isRecord } from "../utils.js";
 import { configIncludeOwnsAgentRosterValues } from "./agent-roster-provenance.js";
-import { applyMergePatch } from "./merge-patch.js";
+import { applyMergePatch, createMergePatch } from "./merge-patch.js";
 import { normalizeAgentModelMapForConfig, normalizeAgentModelRefForConfig } from "./model-input.js";
 import type { OpenClawConfig } from "./types.js";
 
-const OPEN_DM_POLICY_ALLOW_FROM_RE =
-  /^(?<policyPath>[a-z0-9_.-]+)\s*=\s*"open"\s+requires\s+(?<allowPath>[a-z0-9_.-]+)(?:\s+\(or\s+[a-z0-9_.-]+\))?\s+to include "\*"$/i;
-
-const MANAGED_CONFIG_UNSET_PATHS = [["plugins", "installs"]] as const;
 const AGENT_ROSTER_PATHS = [
   ["agents", "entries"],
   ["agents", "list"],
@@ -69,42 +65,6 @@ type ManifestModelIdNormalizationProvider = {
 // Clone config fragments before patching so mutation preparation never aliases callers.
 function cloneUnknown<T>(value: T): T {
   return structuredClone(value);
-}
-
-/** Builds an RFC-7396-style merge patch between source and target config values. */
-export function createMergePatch(base: unknown, target: unknown): unknown {
-  if (!isRecord(base) || !isRecord(target)) {
-    return cloneUnknown(target);
-  }
-
-  const patch: Record<string, unknown> = {};
-  const keys = new Set([...Object.keys(base), ...Object.keys(target)]);
-  for (const key of keys) {
-    const hasBase = Object.hasOwn(base, key);
-    const hasTarget = Object.hasOwn(target, key);
-    if (!hasTarget) {
-      patch[key] = null;
-      continue;
-    }
-    const targetValue = target[key];
-    if (!hasBase) {
-      patch[key] = cloneUnknown(targetValue);
-      continue;
-    }
-    const baseValue = base[key];
-    if (isRecord(baseValue) && isRecord(targetValue)) {
-      const childPatch = createMergePatch(baseValue, targetValue);
-      if (isRecord(childPatch) && Object.keys(childPatch).length === 0) {
-        continue;
-      }
-      patch[key] = childPatch;
-      continue;
-    }
-    if (!isDeepStrictEqual(baseValue, targetValue)) {
-      patch[key] = cloneUnknown(targetValue);
-    }
-  }
-  return patch;
 }
 
 export function projectSourceOntoRuntimeShape(source: unknown, runtime: unknown): unknown {
@@ -1758,306 +1718,11 @@ function preserveRootSchemaUri(params: {
   };
 }
 
-export function formatConfigValidationFailure(pathLabel: string, issueMessage: string): string {
-  const match = issueMessage.match(OPEN_DM_POLICY_ALLOW_FROM_RE);
-  const policyPath = match?.groups?.policyPath?.trim();
-  const allowPath = match?.groups?.allowPath?.trim();
-  if (!policyPath || !allowPath) {
-    return `Config validation failed: ${pathLabel}: ${issueMessage}`;
-  }
-
-  return [
-    `Config validation failed: ${pathLabel}`,
-    "",
-    `Configuration mismatch: ${policyPath} is "open", but ${allowPath} does not include "*".`,
-    "",
-    "Fix with:",
-    `  openclaw config set ${allowPath} '["*"]'`,
-    "",
-    "Or switch policy:",
-    `  openclaw config set ${policyPath} "pairing"`,
-  ].join("\n");
-}
-
 function isNumericPathSegment(raw: string): boolean {
   return parseArrayIndexPathSegment(raw) !== undefined;
 }
 
 function parseArrayIndexPathSegment(raw: string): number | undefined {
   return parseConfigPathArrayIndex(raw);
-}
-
-function isWritePlainObject(value: unknown): value is Record<string, unknown> {
-  return Boolean(value) && typeof value === "object" && !Array.isArray(value);
-}
-
-function hasOwnObjectKey(value: Record<string, unknown>, key: string): boolean {
-  return Object.hasOwn(value, key);
-}
-
-const WRITE_PRUNED_OBJECT = Symbol("write-pruned-object");
-
-function coerceConfig(value: unknown): OpenClawConfig {
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return {};
-  }
-  return value as OpenClawConfig;
-}
-
-function unsetPathForWriteAt(
-  value: unknown,
-  pathSegments: string[],
-  depth: number,
-): { changed: boolean; value: unknown } {
-  if (depth >= pathSegments.length) {
-    return { changed: false, value };
-  }
-  const segment = expectDefined(pathSegments[depth], "path segments entry at depth");
-  const isLeaf = depth === pathSegments.length - 1;
-
-  if (Array.isArray(value)) {
-    const index = parseArrayIndexPathSegment(segment);
-    if (index === undefined || index >= value.length) {
-      return { changed: false, value };
-    }
-    if (isLeaf) {
-      const next = value.slice();
-      next.splice(index, 1);
-      return { changed: true, value: next };
-    }
-    const child = unsetPathForWriteAt(value[index], pathSegments, depth + 1);
-    if (!child.changed) {
-      return { changed: false, value };
-    }
-    const next = value.slice();
-    if (child.value === WRITE_PRUNED_OBJECT) {
-      next.splice(index, 1);
-    } else {
-      next[index] = child.value;
-    }
-    return { changed: true, value: next };
-  }
-
-  if (
-    isBlockedObjectKey(segment) ||
-    !isWritePlainObject(value) ||
-    !hasOwnObjectKey(value, segment)
-  ) {
-    return { changed: false, value };
-  }
-  if (isLeaf) {
-    const next: Record<string, unknown> = { ...value };
-    delete next[segment];
-    return {
-      changed: true,
-      value: Object.keys(next).length === 0 ? WRITE_PRUNED_OBJECT : next,
-    };
-  }
-
-  const child = unsetPathForWriteAt(value[segment], pathSegments, depth + 1);
-  if (!child.changed) {
-    return { changed: false, value };
-  }
-  const next: Record<string, unknown> = { ...value };
-  if (child.value === WRITE_PRUNED_OBJECT) {
-    delete next[segment];
-  } else {
-    next[segment] = child.value;
-  }
-  return {
-    changed: true,
-    value: Object.keys(next).length === 0 ? WRITE_PRUNED_OBJECT : next,
-  };
-}
-
-function unsetPathForWrite(
-  root: OpenClawConfig,
-  pathSegments: string[],
-): { changed: boolean; next: OpenClawConfig } {
-  if (pathSegments.length === 0) {
-    return { changed: false, next: root };
-  }
-  const result = unsetPathForWriteAt(root, pathSegments, 0);
-  if (!result.changed) {
-    return { changed: false, next: root };
-  }
-  if (result.value === WRITE_PRUNED_OBJECT) {
-    return { changed: true, next: {} };
-  }
-  if (isWritePlainObject(result.value)) {
-    return { changed: true, next: coerceConfig(result.value) };
-  }
-  return { changed: false, next: root };
-}
-
-export function applyUnsetPathsForWrite(
-  root: OpenClawConfig,
-  unsetPaths: readonly string[][] | undefined,
-): OpenClawConfig {
-  let next = root;
-  for (const unsetPath of unsetPaths ?? []) {
-    if (!Array.isArray(unsetPath) || unsetPath.length === 0) {
-      continue;
-    }
-    const unsetResult = unsetPathForWrite(next, unsetPath);
-    if (unsetResult.changed) {
-      next = unsetResult.next;
-    }
-  }
-  return next;
-}
-
-export function resolveManagedUnsetPathsForWrite(
-  unsetPaths: readonly string[][] | undefined,
-): string[][] {
-  const next: string[][] = [];
-  for (const managedPath of MANAGED_CONFIG_UNSET_PATHS) {
-    next.push(Array.from(managedPath));
-  }
-  for (const unsetPath of unsetPaths ?? []) {
-    if (!Array.isArray(unsetPath) || unsetPath.length === 0) {
-      continue;
-    }
-    if (next.some((existing) => isDeepStrictEqual(existing, unsetPath))) {
-      continue;
-    }
-    next.push([...unsetPath]);
-  }
-  return next;
-}
-
-export function collectChangedPaths(
-  base: unknown,
-  target: unknown,
-  path: string,
-  output: Set<string>,
-): void {
-  if (Array.isArray(base) && Array.isArray(target)) {
-    const max = Math.max(base.length, target.length);
-    for (let index = 0; index < max; index += 1) {
-      const childPath = path ? `${path}[${index}]` : `[${index}]`;
-      if (index >= base.length || index >= target.length) {
-        output.add(childPath);
-        continue;
-      }
-      collectChangedPaths(base[index], target[index], childPath, output);
-    }
-    return;
-  }
-  if (isRecord(base) && isRecord(target)) {
-    const keys = new Set([...Object.keys(base), ...Object.keys(target)]);
-    for (const key of keys) {
-      const childPath = path ? `${path}.${key}` : key;
-      const hasBase = Object.hasOwn(base, key);
-      const hasTarget = Object.hasOwn(target, key);
-      if (!hasTarget || !hasBase) {
-        output.add(childPath);
-        continue;
-      }
-      collectChangedPaths(base[key], target[key], childPath, output);
-    }
-    return;
-  }
-  if (!isDeepStrictEqual(base, target)) {
-    output.add(path);
-  }
-}
-
-function parentPath(value: string): string {
-  if (!value) {
-    return "";
-  }
-  if (value.endsWith("]")) {
-    const index = value.lastIndexOf("[");
-    return index > 0 ? value.slice(0, index) : "";
-  }
-  const index = value.lastIndexOf(".");
-  return index >= 0 ? value.slice(0, index) : "";
-}
-
-function isPathChanged(path: string, changedPaths: Set<string>): boolean {
-  if (changedPaths.has(path)) {
-    return true;
-  }
-  let current = parentPath(path);
-  while (current) {
-    if (changedPaths.has(current)) {
-      return true;
-    }
-    current = parentPath(current);
-  }
-  return changedPaths.has("");
-}
-
-export function restoreEnvRefsFromMap(
-  value: unknown,
-  path: string,
-  envRefMap: Map<string, string>,
-  changedPaths: Set<string>,
-  identityRestoredPaths: ReadonlySet<string> = new Set(),
-): unknown {
-  if (typeof value === "string") {
-    if (identityRestoredPaths.has(path)) {
-      return value;
-    }
-    if (!isPathChanged(path, changedPaths)) {
-      const original = envRefMap.get(path);
-      if (original !== undefined) {
-        return original;
-      }
-    }
-    return value;
-  }
-  if (Array.isArray(value)) {
-    let changed = false;
-    const next = value.map((item, index) => {
-      const updated = restoreEnvRefsFromMap(
-        item,
-        `${path}[${index}]`,
-        envRefMap,
-        changedPaths,
-        identityRestoredPaths,
-      );
-      if (updated !== item) {
-        changed = true;
-      }
-      return updated;
-    });
-    return changed ? next : value;
-  }
-  if (isRecord(value)) {
-    let changed = false;
-    const next: Record<string, unknown> = {};
-    for (const [key, child] of Object.entries(value)) {
-      const childPath = path ? `${path}.${key}` : key;
-      const updated = restoreEnvRefsFromMap(
-        child,
-        childPath,
-        envRefMap,
-        changedPaths,
-        identityRestoredPaths,
-      );
-      if (updated !== child) {
-        changed = true;
-      }
-      next[key] = updated;
-    }
-    return changed ? next : value;
-  }
-  return value;
-}
-
-export function resolveWriteEnvSnapshotForPath(params: {
-  actualConfigPath: string;
-  expectedConfigPath?: string;
-  envSnapshotForRestore?: Record<string, string | undefined>;
-}): Record<string, string | undefined> | undefined {
-  if (
-    params.expectedConfigPath === undefined ||
-    params.expectedConfigPath === params.actualConfigPath
-  ) {
-    return params.envSnapshotForRestore;
-  }
-  return undefined;
 }
 /* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
