+++ src/system-agent/setup-inference-plan.ts
@@ -10,7 +10,7 @@ import {
   GEMINI_CLI_DEFAULT_MODEL_REF,
   OPENAI_API_DEFAULT_MODEL_REF,
 } from "../commands/onboard-inference.js";
-import { createMergePatch } from "../config/io.write-prepare.js";
+import { createMergePatch } from "../config/merge-patch.js";
 import {
   normalizeAgentModelRefForConfig,
   resolveAgentModelPrimaryValue,
